backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
motor
dnspython
requests
python-multipart
yake
redis>=5.0.0
aioredis>=2.0.1
pyjwt
email-validator
